layout structure example + some console noise cleanup
