+++ app/theme.ts
@@ -245,7 +245,7 @@ const Theme = createTheme({
       sm: 600,
       md: 800,
       lg: 1280,
-      xl: 1920
+      xl: 1500
     }
   }
 })
