+++ package.json
@@ -59,5 +59,6 @@
     "eslint-config-next": "^16.0.3",
     "prettier": "^3.3.3",
     "typescript": "^5"
-  }
+  },
+  "packageManager": "yarn@4.6.0+sha512.5383cc12567a95f1d668fbe762dfe0075c595b4bfff433be478dbbe24e05251a8e8c3eb992a986667c1d53b6c3a9c85b8398c35a960587fbd9fa3a0915406728"
 }
